requirements.txt
bcrypt==3.1.3
cffi==1.10.0
coreapi==2.3.0
coreschema==0.0.4
Django==1.10.6
django-cors-headers==2.0.2
django-filter==1.0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.2
itypes==1.1.0
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.1
passlib==1.7.1
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.4.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
